BAU: Removed method headers
Following discussions with the team regarding code styles it was agreed
to remove the function method headers. The main reason for this was in
regards to maintenance and ensuring engineers are creating clean,
maintainable code without needing to rely upon comments which may not
necessarily be kept up to date.