SPT-1514: Check whether fraud check credential has expired
Initial commit with basic tests to check whether the most recent
CURRENT fraud check credential was issued within the last x hours.
If the fraud check has an `nbf` value outside of this window then
this function will return a value of expired as true.

This will eventually allow IPV Core or Orchestration to know whether
to re-prove the user's identity.

We have chosen to do this in hours because this matches IPV Core's
implementation and the results of the calls to SIS should match those
that IPV Core currently produce.

Co-authored-by: Chaitan Guttarlapalli <chaitan.guttarlapalli@digital.cabinet-office.gov.uk>
Co-authored-by: Muhammed Islam <muhammed.islam@digital.cabinet-office.gov.uk>
Co-authored-by: Michaela Benyohai <michaela.benyohai@digital.cabinet-office.gov.uk>