SPT-1514: Handle international address and fraud resilience scenarios
Addition of logic to check if Fraud VC contains the `failedCheckDetails`
property with the value of `applicable_authoritative_source` or
`available_authoritative_source`. The former implies that the Fraud VC
was written for an international address and the stored identity is
therefore considered valid. The latter implies that the Fraud VC was
written when the fraud check process was temporarily unavailable and
the stored identity is considered invalid (user should return to re-do
a fraud check).

Co-authored-by: Chaitan Guttarlapalli <chaitan.guttarlapalli@digital.cabinet-office.gov.uk>
Co-authored-by: Muhammed Islam <muhammed.islam@digital.cabinet-office.gov.uk>
Co-authored-by: Michaela Benyohai <michaela.benyohai@digital.cabinet-office.gov.uk>