SPT-1635 Silently handle AIS events that are missing intervention_code
Do not throw an error when an AIS event does not contain the
intervention_code field. Instead silently ignore it.

We were expecting an AIS event to always contain an intervention_code
field and throwing an exception when this field was missing. However,
it turns out that this is an optional field and we are receiving AIS
events in production without this field, leading to an error to be
thrown and messages ending up on the DLQ. This commit fixes this issue.