BAU: Fix message processor lambda
In a previous commit the `message-processor.ts` handler code was
renamed `txma-message-processor.ts`. The SAM template was updated to
provide the new `EntryPoint` in the `MetaData` but unfortunately
updating the `Handler` was missed. This lambda is currently throwing
initialisation errors in production.

This commit updates the Handler to be correct and should fix the issue.
Some informational logging has also been added to the handler so that
it is easier to see evidence that it is working when it is working.
Currently when everything is fine no logs are produced.