SPT-1514: Set `expired` on user-identity based on fraud check expiry
Add functionality to the user-identity handler to get the CURRENT
verifiable credentials from storage and pass them to the existing
function in fraudCheckService to work out whether the fraud check has
expired. Set the value of `expired` on the user-identity response - it
is true if the fraud check has expired and false if it has not expired.

As part of this change the `hasFraudCheckExpired` function to directly
consume from `getConfiguration` rather than being passed the vales it
needs from config in its function definition. Configuration is cached
by powertools anyway so there is no benefit to calling it once and
passing it through the function chain.