SPT-1514: Update acceptance test for case with 0 credentials
The current acceptance tests are partially implemented and do not store
credentials in EVCS. The acceptance tests could assert that the
identity was returned from /user-identity with `expired` as true.
Now that we have implemented the logic to look for a valid fraud
credential, a bundle of zero credentials will cause the logic to set
`expired` to 0. This commit therefore makes the acceptance tests expect
a `expired` value of true to be returned. In a future PR we will extend
the acceptance tests to send and return credentials, allowing us to
ully test this logic. This is a temporary measure to keep the current
PR small for review.