SPT-1629: Reframe acceptance tests with 0 identity checks as "unhappy"
We currently have an acceptance test that sets up a scenario with a
stored identity and zero associated credentials which lives in the
"happy path" feature file. However any scenario with zero credentials
is likely to be unhappy and marked with `isValid=false` given the
credentials don't match the stored identity.

This also ensures any "unhappy" path tests are defined to contain
credentials, unless they are explicitly testing the zero credential
scenario. This it to ensure they are not accidentally testing the
wrong thing.

These tests will fail until the implementation is completed (TDD).