SPT-1629: Refactor acceptance tests to send credentials to EVCS
The acceptance tests currently put stored identity records into EVCS
as part of test set up, however they do not put the corresponding
credentials into EVCS as these were previously not needed for the
use cases we were testing. We will soon be implementing logic that
consumes the individual credentials as well as the stored identity.
Therefore this commit introduces steps to the existing acceptance tests
that also puts the credentials into EVCS.

Co-authored-by: Michaela Benyohai <michaela.benyohai@digital.cabinet-office.gov.uk>
Co-authored-by: Umar Hansa <umar.hansa@digital.cabinet-office.gov.uk>