SPT-1629: Add acceptance tests for validating the credentials
We want to make sure that the stored identity record references the
signatures for each individual credential that was used to calculate
the vot for the user. If it references different credentials to those
that are returned from EVCS it should set the value of isValid to false
on the /user-identity endpoint. These tests validate that behaviour.

Co-authored-by: Michaela Benyohai <michaela.benyohai@digital.cabinet-office.gov.uk>
Co-authored-by: Umar Hansa <umar.hansa@digital.cabinet-office.gov.uk>