SPT-1629: Add `credentials` property to user-identity response schema
The openAPI schema for the user-identity response does not contain the
`credentials` property. This will be returned for phase 2, whilst IPV
Core is consuming this endpoint. For now the endpoint will not return
`https://vocab.account.gov.uk/v1/credentialJWT` in the response. This
will be added in phase 3 when Orchestration consumes the endpoint.