SPT-1719: Fix vot field on SIS_STORED_IDENTITY_RETURNED event
The vot field on the SIS_STORED_IDENTITY_RETURNED is supposed to be the
value of the calculated vot, so dependent on the vtr requested. At the
moment this field is set to the maximum vot recorded by IPV Core, which
is not always equal to the calculated vot.

This commit updates the field to report the calculated vot. The unit
tests for user-identity-handler have been updated to use a scenario
where the max vot and requested vtr are different (P3 and P2) so make
sure we are testing this scenario effectively. Previously the scenario
being tested set the vtr and max vot as both to P2, meaning that we
were not catching bugs when testing the audit messages.

It also emerged that unit test for the audit message was testing that
the response_body field qas equal to the stored identity JWT returned
from EVCS. However this was incorrect, and it should be set to the
value of the response body from the /user-identity endpoint on SIS.
This aspect of the test has also been fixed.