SPT-1520: Fix Typescript type error
There is a type error in user-identityresponse.ts that we did not spot
earlier. We are trying to assign the string "P0" to a generic type
within StoredIdentityJWT that extends from `IdentityVectorOfTrust`.
However, "P0" does not extend from IdentityVectorOfTrust, hence the
error. This commit fixes the issue by making the generic type extend
string, which "P0" is.